Fix suggestions container positioning in relative positioned containers
The suggestions container is normally/currently positioned directly
after the text field element (inside the same parent container).

It is absolutely positioned and we aim to have it appear beneath
the text field.
Absolute positioning works differently for elements which are inside
`position: relative` containers. The top/left coordinates are considered
relative to the relative container element, not to the whole document.

`$.offset()` gives us document-based coordinates, which don't work in
this case.
`$.position()` should be used instead.

Another solution would be to put the suggestions container
elsewhere in the DOM, not after the text field. Perhaps
at the end of the document body.